requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.21.2
botocore==1.24.2
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
idna==3.3
jmespath==0.10.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
s3transfer==0.5.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
PyQRCode==1.2.1
PyJWT==2.10.1
django-otp==1.5.4
